offensive-masterlist.txt
List of words deemed offensive, nonsensical, or potentially offensive when paired with other words (e.g. country names). Used to 'clean' passphrase dictionaries.

List includes words from a number of languages.  It is surprisingly comprehensive, but creative sailors are welcome to suggest additions! Please just check that your contributions are not already in the list.